Initial checkin of the zero copy, lockless ring buffer
This ring buffer implementation is for a fixed slot ring buffer
that is designed for single reader single writer usage. this
allows it to be lockless. It also allows for zero copy semantics
by providing a two stage read/write mechanism that provides direct
access to the underlying ring buffer slots.